--- bdml/doc.go
@@ -1,2 +1,0 @@
-// Package bdml handles encoding/decoding of the Birdie Data Manipulation Language.
-package bdml
