+++ dml/dml.go
@@ -1,4 +1,4 @@
-package bdml
+package dml
 
 import "unique"
