+++ dml/doc.go
@@ -1,0 +1,2 @@
+// Package dml handles encoding/decoding of the Data Manipulation Language.
+package dml
